Fix ICC change-of-sign warning
INT_MIN is negative.

qv4assembler_p.h(108): warning #68: integer conversion resulted in a
change of sign

+++ src/qml/jit/qv4assembler_p.h
@@ -105,7 +105,7 @@ struct LookupCall {
 struct RuntimeCall {
     JSC::MacroAssembler::Address addr;
 
-    inline RuntimeCall(uint offset = INT_MIN);
+    inline RuntimeCall(uint offset = uint(INT_MIN));
     bool isValid() const { return addr.offset >= 0; }
 };
